requirements.txt
streamlit==1.23.1
pandas==1.5.3
numpy==1.24.3
google-api-core==2.11.0
google-cloud-bigquery==3.5.0
langchain==0.0.176
openai==0.27.8
seaborn==0.12.2
matplotlib==3.7.1
db-dtypes==1.0.5
streamlit-option-menu==0.3.6
pandasql==0.7.3
pyodbc==4.0.34
pypyodbc==1.3.6
SQLAlchemy==1.4.39
databricks-api==0.9.0
PyQt5==5.15.9
